Schema Registry (#6)
* Create Project - Schema Registry

* Functional Schema Registry

For now it saves the schemas in a json file. The compatibility checker always returns true. Apart from that the schema registry works, endpoints for saving and downloading schemas work and the project structure is more or less prepared.

* Compatibility Check structure

Change the types in the interface and adjust the SchemaRegistryService's IsComatible function.

* Compatibility Checker implementation

Implemented a compatibility checker. For now it doesn't accept changing field names.

* Fix a mistake

* SchemaEquals

* Tests

* Tests for SchemaRegistry

* Changes Requested in PR comments

* Requested corrections

* Change the file structure a bit

As requested

* Hexagonal architecture

* Update Tests

* cleanup

* New Schema Validation

and test update

* Divide Program.cs

* Primary constructor

* Requested changes

* Requested adjustments